Add option to hide FPS slider via spawn.ini setting
WIP: First time doing something like this from scratch. Not simple rename. Be very cautious on merging / approving.

Introduces a new HideFPSSlider setting in spawn.ini to control the visibility of the FPS/gamespeed slider in the in-game pause menu.

Updates include a new ASM patch to conditionally hide the slider, Makefile changes to include the patch, and documentation for usage.

This is useful for tournament play or restricting in-game speed adjustments.